refactor: prettier

+++ packages/smart-contracts-boilerplate/contracts/Domi.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.0;
 import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
 import './Ownable.sol';
 import 'abdk-libraries-solidity/ABDKMath64x64.sol';
+
 abstract contract PrincipalInterface {
   function distributeSavingsRate(uint256 amount) external virtual;
 }
@@ -106,18 +107,20 @@ contract DomiToken is ERC20, Ownable {
     }
   }
 
-  function getSavingsRate() public view returns (uint) {
+  function getSavingsRate() public view returns (uint256) {
     return savingsRate;
   }
-  function setSavingsRate(uint _savingsRate) public {
-    savingsRate=_savingsRate;
+
+  function setSavingsRate(uint256 _savingsRate) public {
+    savingsRate = _savingsRate;
   }
-  function getDomiPrice(uint idx) public pure returns (uint) {
+
+  function getDomiPrice(uint256 idx) public pure returns (uint256) {
     uint16[] memory price = new uint16[](3);
     price[0] = 10000;
     price[1] = 9000;
     price[2] = 12000;
-    uint domiPrice=uint256(price[idx]);
+    uint256 domiPrice = uint256(price[idx]);
     return domiPrice;
   }
 }
